First working version
This upload has the first working version of the script that trains the model. The training takes many hours and is still in progress. The code has been troubleshot and will successfully train a model. Components should be arranged in the right folders as will be provided in a ReadMe file.